requirements.txt
torch>=2.1.0
torchvision
numpy
opencv-python
Pillow
flask
tqdm
rembg
onnxruntime-gpu
flask-cors
segment-anything
tqdm>=4.64.0